+++ src/ui_worm.cpp
@@ -1,4 +1,3 @@
-//ui_worm.cpp
 #include <Geode/Geode.hpp>
 #include <Geode/modify/MenuLayer.hpp>
 #include <Geode/binding/CCMenuItemSpriteExtra.hpp>
@@ -19,8 +18,36 @@ V  \ \/  /      \       /
 using namespace geode::prelude;
 using namespace cocos2d;
 
-//default setting, no worm :(
-bool gWormOn = false;
+//0 = off, 1 = worm, 2 = big worm
+int  gWormMode = 0;
+int  gWormHistory = 90;
+
+static const char* modeLabel(int mode) {
+    switch (mode) {
+    case 1:  return "worm";
+    case 2:  return "big worm";
+    default: return "no worm";
+    }
+}
+
+static ccColor3B modeTint(int mode) {
+    switch (mode) {
+    case 1:  return { 255, 255, 255 };
+    case 2:  return { 255, 220, 160 };
+    default: return { 180, 180, 180 };
+    }
+}
+
+static void applyHistoryForMode() {
+    //normal worm is only off screen at 4x speed
+    if (gWormMode == 1)
+        gWormHistory = 90;
+    //big worm goes off screen even at 0.5x speed
+    else if (gWormMode == 2)
+        gWormHistory = 200;
+    else
+        gWormHistory = 90;
+}
 
 class $modify(WormToggleMenu, MenuLayer) {
 public:
@@ -48,10 +75,15 @@ public:
         if (auto* s = CCSprite::create("button.png"_spr)) {
             base = s;
         }
+        else {
+            base = CCSprite::create();
+            base->setContentSize({ 120.f, 48.f });
+        }
+        applyHistoryForMode();
 
         //worm button label
-        auto* label = CCLabelBMFont::create(gWormOn ? "worm" : "no worm", "bigFont.fnt");
-        label->setScale(0.3f);
+        auto* label = CCLabelBMFont::create(modeLabel(gWormMode), "bigFont.fnt");
+        label->setScale(0.27f);
         label->setPosition({
             base->getContentSize().width / 2.f,
             base->getContentSize().height / 2.f
@@ -61,7 +93,7 @@ public:
         m_fields->label = label;
         m_fields->buttonSprite = base;
 
-        //worm button clickable
+        //make worm button clickable
         auto* item = CCMenuItemSpriteExtra::create(base, this, menu_selector(WormToggleMenu::onWormButton));
         item->setID("worm-toggle-button"_spr);
 
@@ -69,20 +101,24 @@ public:
         bottomMenu->addChild(item);
         bottomMenu->updateLayout();
 
-        //worm button tint for on/off
+        //set tint depending on current mode
         if (m_fields->buttonSprite)
-            m_fields->buttonSprite->setColor(gWormOn ? ccColor3B{ 255,255,255 } : ccColor3B{ 180,180,180 });
+            m_fields->buttonSprite->setColor(modeTint(gWormMode));
 
         return true;
     }
 
+    //toggle between no worm/worm/big worm
     void onWormButton(CCObject*) {
-        gWormOn = !gWormOn;
+        gWormMode = (gWormMode + 1) % 3;
+        applyHistoryForMode();
 
+        //update label text
         if (m_fields->label)
-            m_fields->label->setString(gWormOn ? "worm" : "no worm");
+            m_fields->label->setString(modeLabel(gWormMode));
 
+        //update button tint
         if (m_fields->buttonSprite)
-            m_fields->buttonSprite->setColor(gWormOn ? ccColor3B{ 255,255,255 } : ccColor3B{ 180,180,180 });
+            m_fields->buttonSprite->setColor(modeTint(gWormMode));
     }
 };