+++ changelog.md
@@ -1,2 +1,8 @@
-# 1.0.0
-- Edit this file to change your mod's changelog.
+# Changelog
+
+## 1.0.10
+- Added Changelog
+- Replaced CCARRAY_FOREACH with CCArrayExt
+- Use typeinfo_cast instead of dynamic_cast
+- Changed normal worm mode length
+- Added big worm mode